+++ tap-snapshots/lib-test-TAP.test.js
@@ -6,99 +6,159 @@
  */
 'use strict'
 exports[`lib/test TAP Query SSR with fetch unavailable. > GraphQL request cache. 1`] = `
-{ fetchError: 'Global fetch API or polyfill unavailable.' }
+{
+  "fetchError": "Global fetch API or polyfill unavailable."
+}
 `
 
 exports[`lib/test TAP Query SSR with fetch unavailable. > Query render function arguments. 1`] = `
-[Arguments] {
-  '0':
-   { load: [Function],
-     loading: true,
-     fetchError: 'Global fetch API or polyfill unavailable.' } }
+{
+  "0": {
+    "load": "[Function]",
+    "loading": true,
+    "fetchError": "Global fetch API or polyfill unavailable."
+  }
+}
 `
 
 exports[`lib/test TAP Query SSR with relative fetch URL. > GraphQL request cache. 1`] = `
-{ fetchError: 'Only absolute URLs are supported' }
+{
+  "fetchError": "Only absolute URLs are supported"
+}
 `
 
 exports[`lib/test TAP Query SSR with relative fetch URL. > Query render function arguments. 1`] = `
-[Arguments] {
-  '0':
-   { load: [Function],
-     loading: true,
-     fetchError: 'Only absolute URLs are supported' } }
+{
+  "0": {
+    "load": "[Function]",
+    "loading": true,
+    "fetchError": "Only absolute URLs are supported"
+  }
+}
 `
 
 exports[`lib/test TAP Query SSR with HTTP error. > GraphQL request cache. 1`] = `
-{ httpError: { status: 404, statusText: 'Not Found' },
-  parseError:
-   'invalid json response body at http://localhost:<port>/ reason: Unexpected token N in JSON at position 0' }
+{
+  "httpError": {
+    "status": 404,
+    "statusText": "Not Found"
+  },
+  "parseError": "invalid json response body at http://localhost:<port>/ reason: Unexpected token N in JSON at position 0"
+}
 `
 
 exports[`lib/test TAP Query SSR with HTTP error. > Query render function arguments. 1`] = `
-[Arguments] {
-  '0':
-   { load: [Function],
-     loading: true,
-     httpError: { status: 404, statusText: 'Not Found' },
-     parseError:
-      'invalid json response body at http://localhost:<port>/ reason: Unexpected token N in JSON at position 0' } }
+{
+  "0": {
+    "load": "[Function]",
+    "loading": true,
+    "httpError": {
+      "status": 404,
+      "statusText": "Not Found"
+    },
+    "parseError": "invalid json response body at http://localhost:<port>/ reason: Unexpected token N in JSON at position 0"
+  }
+}
 `
 
 exports[`lib/test TAP Query SSR with response JSON invalid. > GraphQL request cache. 1`] = `
-{ parseError:
-   'invalid json response body at http://localhost:<port>/ reason: Unexpected token N in JSON at position 0' }
+{
+  "parseError": "invalid json response body at http://localhost:<port>/ reason: Unexpected token N in JSON at position 0"
+}
 `
 
 exports[`lib/test TAP Query SSR with response JSON invalid. > Query render function arguments. 1`] = `
-[Arguments] {
-  '0':
-   { load: [Function],
-     loading: true,
-     parseError:
-      'invalid json response body at http://localhost:<port>/ reason: Unexpected token N in JSON at position 0' } }
+{
+  "0": {
+    "load": "[Function]",
+    "loading": true,
+    "parseError": "invalid json response body at http://localhost:<port>/ reason: Unexpected token N in JSON at position 0"
+  }
+}
 `
 
 exports[`lib/test TAP Query SSR with response payload malformed. > GraphQL request cache. 1`] = `
-{ parseError: 'Malformed payload.' }
+{
+  "parseError": "Malformed payload."
+}
 `
 
 exports[`lib/test TAP Query SSR with response payload malformed. > Query render function arguments. 1`] = `
-[Arguments] {
-  '0':
-   { load: [Function],
-     loading: true,
-     parseError: 'Malformed payload.' } }
+{
+  "0": {
+    "load": "[Function]",
+    "loading": true,
+    "parseError": "Malformed payload."
+  }
+}
 `
 
 exports[`lib/test TAP Query SSR with GraphQL errors. > GraphQL request cache. 1`] = `
-{ httpError: { status: 400, statusText: 'Bad Request' },
-  graphQLErrors:
-   [ { message: 'Syntax Error: Unexpected Name "x"',
-       locations: [ { line: 1, column: 1 } ] } ] }
+{
+  "httpError": {
+    "status": 400,
+    "statusText": "Bad Request"
+  },
+  "graphQLErrors": [
+    {
+      "message": "GraphQL query syntax error: Syntax Error: Unexpected Name \\"x\\"",
+      "locations": [
+        {
+          line: 1,
+          column: 1
+        }
+      ]
+    }
+  ]
+}
 `
 
 exports[`lib/test TAP Query SSR with GraphQL errors. > Query render function arguments. 1`] = `
-[Arguments] {
-  '0':
-   { load: [Function],
-     loading: true,
-     httpError: { status: 400, statusText: 'Bad Request' },
-     graphQLErrors:
-      [ { message: 'Syntax Error: Unexpected Name "x"',
-          locations: [ { line: 1, column: 1 } ] } ] } }
+{
+  "0": {
+    "load": "[Function]",
+    "loading": true,
+    "httpError": {
+      "status": 400,
+      "statusText": "Bad Request"
+    },
+    "graphQLErrors": [
+      {
+        "message": "GraphQL query syntax error: Syntax Error: Unexpected Name \\"x\\"",
+        "locations": [
+          {
+            line: 1,
+            column: 1
+          }
+        ]
+      }
+    ]
+  }
+}
 `
 
 exports[`lib/test TAP Query SSR with variables. > GraphQL request cache. 1`] = `
-{ data: { date: { year: 2018 } } }
+{
+  "data": {
+    "date": {
+      "year": 2018
+    }
+  }
+}
 `
 
 exports[`lib/test TAP Query SSR with variables. > Query render function arguments. 1`] = `
-[Arguments] {
-  '0':
-   { load: [Function],
-     loading: true,
-     data: { date: { year: 2018 } } } }
+{
+  "0": {
+    "load": "[Function]",
+    "loading": true,
+    "data": {
+      "date": {
+        "year": 2018
+      }
+    }
+  }
+}
 `
 
 exports[`lib/test TAP Query SSR with nested query. > HTML displaying the nested query render function argument. 1`] = `